+++ truffle/src/com.oracle.truffle.polyglot/src/com/oracle/truffle/polyglot/PolyglotExceptionImpl.java
@@ -108,7 +108,7 @@ final class PolyglotExceptionImpl extends AbstractExceptionImpl implements com.o
         this.engine = engine;
         this.context = (languageContext != null) ? languageContext.context : null;
         this.exception = original;
-        this.guestFrames = TruffleStackTrace.getStacktrace(original);
+        this.guestFrames = TruffleStackTrace.getStackTrace(original);
 
         if (exception instanceof TruffleException) {
             TruffleException truffleException = (TruffleException) exception;